requirements.txt
Django==5.2.9
django-unfold==0.75.0